Avoid undefined shift in endianness conversion

+++ celt/opus_custom_demo.c
@@ -320,7 +320,7 @@ int main(int argc, char *argv[])
             for(i=0;i<frame_size*channels;i++)
             {
                float_bits s;
-               s.i=fbytes[4*i+3]<<24|fbytes[4*i+2]<<16|fbytes[4*i+1]<<8|fbytes[4*i];
+               s.i=(unsigned)fbytes[4*i+3]<<24|fbytes[4*i+2]<<16|fbytes[4*i+1]<<8|fbytes[4*i];
                in[i]=(int)floor(.5 + s.f*8388608);
             }
          }
